# src/Ims.Case.Api/wwwroot/docs/case-v1p0-openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Competencies and Academic Standards Exchange (CASE) Service OpenAPI (JSON) Definition",
    "description": "The Competencies and Academic Standards Exchange (CASE) Service enables the exchange of data between a Competency Records Service Provider and the consumers of the associated data. This service has been described using the IMS Model Driven Specification development approach this being the Platform Specific Model (PSM) of the service.",
    "termsOfService": "http://www.opendefinition.org/licenses/odc-by",
    "contact": {
      "name": "Brad R. Marshall ",
      "email": "bmarshall@sped.mobi"
    },
    "license": {
      "name": "Open Data Commons Attribution License (ODCAL)",
      "url": "http://www.opendefinition.org/licenses/odc-by"
    },
    "version": "v1p0"
  },
  "servers": [
    {
      "url": "https://teks-api.texasgateway.org/ims/case/v1p0",
      "description": "TEA Texas Gateway Server"
    }
  ],
...
  "paths": {
...
    "/CFItemAssociations/{sourcedId}": {
      "get": {
        "tags": [
          "AssociationsManager"
        ],
        "summary": "The REST read request message for the getCFItemAssociations() API call.",
        "description": "This is a request to the Service Provider to provide the all of the Competency Associations for the specified CFItem and the information about the CFItem itself.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
        "operationId": "getCFItemAssociations",
        "parameters": [
          {
            "name": "sourcedId",
            "in": "path",
            "description": "The UUID that identifies the CFItem for which ALL associations are to be supplied.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This is the response when the request has been completed successfully. It is the CFItem and set of CFAssociations from the service provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFAssociationSet"
                }
              }
            }
          },
          "400": {
            "description": "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "403": {
            "description": "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "429": {
            "description": "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "500": {
            "description": "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "default": {
            "description": "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          }
        }
      }
    },
    "/CFItemTypes/{sourcedId}": {
      "get": {
        "tags": [
          "DefinitionsManager"
        ],
        "summary": "The REST read request message for the getCFItemType() API call.",
        "description": "This is a request to the Service Provider to provide the specified Competency Framework Item Type and the set of children CFItemTypes as identified by the hierarchy codes.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
        "operationId": "getCFItemType",
        "parameters": [
          {
            "name": "sourcedId",
            "in": "path",
            "description": "The UUID that identifies the Competency Framework ItemType that is to be read from the service provider.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This is the response when the request has been completed successfully. It is the CFItemType and child CFItemTypes from the service provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFItemTypeSet"
                }
              }
            }
          },
          "400": {
            "description": "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "403": {
            "description": "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "429": {
            "description": "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "500": {
            "description": "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "default": {
            "description": "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          }
        }
      }
    },
    "/CFItems/{sourcedId}": {
      "get": {
        "tags": [
          "ItemsManager"
        ],
        "summary": "The REST read request message for the getCFItem() API call.",
        "description": "This is a request to the Service Provider to provide the specified Competency Framework Item.  If the identified record cannot be found then the 'unknownobject' status code must be reported.",
        "operationId": "getCFItem",
        "parameters": [
          {
            "name": "sourcedId",
            "in": "path",
            "description": "The UUID that identifies the Competency Framework Item that is to be read from the service provider.",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "This is the response when the request has been completed successfully. It is the CFItem from the service provider.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CFItem"
                }
              }
            }
          },
          "400": {
            "description": "An invalid selection field was supplied and data filtering on the selection criteria was not possible i.e. 'invalid_selection_field'. This is accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "401": {
            "description": "The request was not correctly authorised i.e. 'unauthorisedrequest'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "403": {
            "description": "This is used to indicate that the server can be reached and process the request but refuses to take any further action i.e. 'forbidden'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "404": {
            "description": "Either the supplied identifier is unknown in the Service Provider and so the object could not be changed or an invalid UUID has been supplied. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'. The statement 'Unknown Object' of 'Invalid UUID' should also be presented.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "429": {
            "description": "The server is receiving too many requests i.e. 'server_busy'. Retry at a later time. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "500": {
            "description": "This code should be used only if there is catastrophic error and there is not a more appropriate code i.e. 'internal_server_error'. This would be accompanied by the 'codeMajor/severity' values of 'failure/error'.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          },
          "default": {
            "description": "This is the response data payload to be supplied when the HTTP code is NOT explicitly defined. This would be accompanied by the 'codeMajor/severity' values of 'failure/error' and the appropriate 'codeMinor' value. The associated HTTP code will also be supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/imsx_StatusInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CFAssociationSet": {
        "required": [
          "CFAssociations",
          "CFItem"
        ],
        "type": "object",
        "properties": {
          "CFItem": {
            "$ref": "#/components/schemas/CFItem"
          },
          "CFAssociations": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CFPckgAssociation"
            }
          }
        },
        "additionalProperties": false,
        "description": "This is the container for a collection of CFAssociations. There must be at least one CFAssociation. Note that the association can be between CFDocuments or between CFItems."
      },
...
      "CFItem": {
        "type": "object",
        "description": "This is the container for the CFItem data outside of the context of a CFPackage. This is the content that either describes a specific competency (learning objective) or describes a grouping of competencies within the taxonomy of a Competency Framework Document. ",
        "allOf": [
          {
            "$ref": "#/components/schemas/CFPckgItem"
          },
          {
            "required": [
              "CFDocumentURI"
            ],
            "type": "object",
            "properties": {
              "CFDocumentURI": {
                "$ref": "#/components/schemas/LinkURI"
              }
            },
            "additionalProperties": false
          }
        ]
      },
      "CFItemType": {
        "required": [
          "description",
          "hierarchyCode",
          "identifier",
          "lastChangeDateTime",
          "title",
          "uri"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          },
          "title": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "description": {
            "type": "string",
            "description": "Model Primitive Datatype = String"
          },
          "hierarchyCode": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "typeCode": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "lastChangeDateTime": {
            "type": "string",
            "description": "Model Primitive Datatype = DateTime",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "The container for the ItemType information use within the competency framework."
      },
      "CFItemTypeSet": {
        "required": [
          "CFItemTypes"
        ],
        "type": "object",
        "properties": {
          "CFItemTypes": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CFItemType"
            }
          }
        },
        "additionalProperties": false,
        "description": "The container for the set of CFItemTypes supplied in the response payload. The relationship between the CFItemTypes is determined by the 'hierarchyCode'. The first CFItemType is that which has been specified in the call. The other CFItemTypes are the set of children as determined by their place in the 'hierarchyCode' of the CFItemType."
      },
...
      "CFPckgAssociation": {
        "required": [
          "associationType",
          "destinationNodeURI",
          "identifier",
          "lastChangeDateTime",
          "originNodeURI",
          "uri"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "associationType": {
            "type": "string",
            "enum": [
              "isChildOf",
              "isPeerOf",
              "isPartOf",
              "exactMatchOf",
              "precedes",
              "isRelatedTo",
              "replacedBy",
              "exemplar",
              "hasSkillLevel"
            ]
          },
          "sequenceNumber": {
            "type": "integer",
            "description": "Model Primitive Datatype = Integer",
            "format": "int32"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          },
          "originNodeURI": {
            "$ref": "#/components/schemas/LinkGenURI"
          },
          "destinationNodeURI": {
            "$ref": "#/components/schemas/LinkGenURI"
          },
          "CFAssociationGroupingURI": {
            "$ref": "#/components/schemas/LinkURI"
          },
          "lastChangeDateTime": {
            "type": "string",
            "description": "Model Primitive Datatype = DateTime",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the data about the relationship between two CFDocuments or between two CFItems within the context of a CFPackage."
      },
...
      "CFPckgItem": {
        "required": [
          "fullStatement",
          "identifier",
          "lastChangeDateTime",
          "uri"
        ],
        "type": "object",
        "properties": {
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "fullStatement": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "alternativeLabel": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "CFItemType": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          },
          "humanCodingScheme": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "listEnumeration": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "abbreviatedStatement": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "conceptKeywords": {
            "minItems": 0,
            "type": "array",
            "description": "Model Primitive Datatype = NormalizedString",
            "items": {
              "type": "string"
            }
          },
          "conceptKeywordsURI": {
            "$ref": "#/components/schemas/LinkURI"
          },
          "notes": {
            "type": "string",
            "description": "Model Primitive Datatype = String"
          },
          "language": {
            "type": "string",
            "description": "Model Primitive Datatype = Language"
          },
          "educationLevel": {
            "minItems": 0,
            "type": "array",
            "description": "Model Primitive Datatype = NormalizedString",
            "items": {
              "type": "string"
            }
          },
          "CFItemTypeURI": {
            "$ref": "#/components/schemas/LinkURI"
          },
          "licenseURI": {
            "$ref": "#/components/schemas/LinkURI"
          },
          "statusStartDate": {
            "type": "string",
            "description": "Model Primitive Datatype = Date",
            "format": "date"
          },
          "statusEndDate": {
            "type": "string",
            "description": "Model Primitive Datatype = Date",
            "format": "date"
          },
          "lastChangeDateTime": {
            "type": "string",
            "description": "Model Primitive Datatype = DateTime",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the CFItem data within the context of a CFPackage. This is the content that either describes a specific competency (learning objective) or describes a grouping of competencies within the taxonomy of a Competency Framework Document. "
      },
...
      "LinkURI": {
        "required": [
          "identifier",
          "title",
          "uri"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Model Primitive Datatype = NormalizedString"
          },
          "identifier": {
            "$ref": "#/components/schemas/UUID"
          },
          "uri": {
            "type": "string",
            "description": "Model Primitive Datatype = AnyURI",
            "format": "uri"
          }
        },
        "additionalProperties": false,
        "description": "A container for the information that is used to achieve the link data reference."
      },
...
      "UUID": {
        "pattern": "[0-9a-f]{8}-[0-9a-f]{4}-[1-5]{1}[0-9a-f]{3}-[8-9a-b]{1}[0-9a-f]{3}-[0-9a-f]{12}",
        "type": "string",
        "description": "The data-type for establishing a Globally Unique Identifier (GUID). The form of the GUID is a Universally Unique Identifier (UUID) of 16 hexadecimal characters (lower case) in the format 8-4-4-4-12. All permitted versions (1-5) and variants (1-2) are supported.",
        "format": "guid"
      },
      "imsx_CodeMinor": {
        "required": [
          "imsx_codeMinorField"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMinorField": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/imsx_CodeMinorField"
            }
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the set of code minor status codes reported in the responses from the Service Provider."
      },
...
      "imsx_StatusInfo": {
        "required": [
          "imsx_codeMajor",
          "imsx_severity"
        ],
        "type": "object",
        "properties": {
          "imsx_codeMajor": {
            "type": "string",
            "enum": [
              "success",
              "processing",
              "failure",
              "unsupported"
            ]
          },
          "imsx_severity": {
            "type": "string",
            "enum": [
              "status",
              "warning",
              "error"
            ]
          },
          "imsx_description": {
            "type": "string",
            "description": "Model Primitive Datatype = String"
          },
          "imsx_codeMinor": {
            "$ref": "#/components/schemas/imsx_CodeMinor"
          }
        },
        "additionalProperties": false,
        "description": "This is the container for the status code and associated information returned within the HTTP messages received from the Service Provider. For the CASE service this object will only be returned to provide information about a failed request i.e. it will NOT be in the payload for a successful request. See Appendix B for further information on the interpretation of the information contained within this class"
      }
    }
  }
}